refactor(sync): extract sync_engine/lwd.rs
Move the gRPC transport and compact-block / subtree-root download
helpers out of sync_engine/mod.rs into a dedicated lwd submodule:

  - LwdTorGuard type alias
  - open_lwd_channel (plain-TLS / Tor-isolated branch)
  - download_subtree_roots (sapling + orchard)
  - download_blocks (into MemoryBlockSource)

`open_lwd_channel` stays `pub(crate)` (re-exported from mod.rs) because
api/wallet.rs and wallet/sync.rs both call it directly for the
chain-tip probe and for TX broadcast / fee probing. The subtree
roots and block download helpers are internal to the sync loop and
use plain `pub(super)` visibility via `use lwd::{...}` in mod.rs.

As part of the re-export cleanup, tighten the error re-exports: only
`SyncError` is used outside the crate's sync_engine module (by
wallet/tor.rs); `RecoveryStrategy`, `REWIND_DISTANCE`, and
`MAX_REWINDS_PER_RUN` are now plain `use` imports in mod.rs (or
internal to error.rs) rather than `pub(crate)`.

`open_db` and `is_sqlite_lock_contention` stay in mod.rs for now —
they are used only by the orchestration loop (`run_sync_impl`) and
don't belong in lwd. The `sqlite_lock_contention_is_recognised` test
stays with them.

mod.rs is now 690 lines (from 909 after R.3 / 1205 before the
refactor began). 23 tests pass.
